Remove user-supplied input from filename
This changes the name of the files saved to disk from having a random
name like:

    a0b63f9dad0aa2efee76f75956d0f554d5efa9b93fb6b1ae2c1d40a7bb4a98e3.csv.gpg

to a name like:

    2018-02-01_100500.csv.gpg

This fixes a security vulnerability that would allow a malicious user to
write to control where a file is written to disk.

It also makes the files easier to read.

fixes #1